Moves files around into two modules: core and game
The "core" module is intended for engine internals that can be used by
files in "game" to interact with the engine.

The "game" module is for game-specific logic.

This is a major enough change to warrant a minor version increment.
Welcome to 0.2.1!
