+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "project_peril"
-version = "0.2.0"
+version = "0.2.1"
 authors = ["Lars-Ivar Hesselberg Simonsen <larsivarhs@gmail.com>","Matthew Guise <reaverrikku@gmail.com>"]
 build = "build.rs"
