+++ src/renderer/mod.rs
@@ -6,6 +6,7 @@ use ash::util::Align;
 use ash::version::{DeviceV1_0, EntryV1_0, InstanceV1_0};
 use ash::vk;
 use ash::{Device, Entry, Instance};
+use core::Config;
 use image;
 use std::ffi::{CStr, CString};
 use std::fs::File;
@@ -24,8 +25,6 @@ mod presentpass;
 
 pub use self::mainpass::MainPass;
 pub use self::presentpass::PresentPass;
-
-use config::Config;
 
 pub struct Texture
 {
